Add new type of problem to slide under Fix bugs with skill activation connected to float int convertation

+++ Assets/Script/GamePlay/Skills/ISkillHero.cs
@@ -8,7 +8,7 @@ namespace Assets.Script.GamePlay.Skills
 	{
 		bool IsActive { get; }
 		 int SkillValue { get; }
-		bool ActivateSkill(float power,Hero hero,SHeroAnimator heroAnimator);
+		bool ActivateSkill(int power,Hero hero,SHeroAnimator heroAnimator);
 		void SetSkill();
 		void Froze();
 		void Unfroze();
